Add first test script that runs Uscf.x for LiCoO2
Test script that loads a LiCoO2 structure from a cif file
taken from COD. The primitive cell is obtained by using seekpath
which makes that a requirement for it to run, along with pymatgen.

Note that this example will not run with AiiDA v0.9.0 as there is
a bug in the new pluginloader that should be hotfixed soon.

On top of that I had to hack the Quantum ESPRESSO plugin in the
aiida_core tree to sort the atomic species alphabetically such that
Cobalt was listed first, otherwise Uscf.x would crash. Uscf.x
requires that the atoms with a Hubbard U parameter are listed first
in the ATOMIC_POSITIONS card generated for the input file of the
pw.x calculation. Currently there is no way to control the output
order for the QE plugin input writers.