Add option to calculation test script to provide structure node from command line
Currently, the Uscf code requires a specific ordering of the entries in the
ATOMIC_POSITIONS card of the input file that cannot be controlled yet through
the AiidA Quantum ESPRESSO plugin. The order will be determined by which order
the atoms were appended to the StructureData. The script provides a function
to construct a structure node from an included cif file, but there the order
of the atoms will be incorrect and hence the Uscf calculation will fail. Until
this is fixed either in Uscf.x or the Quantum ESPRESSO plugin provides a way
to control the order in the generated input file, we provide the command line
option to specify a StructureData node that will result in a correct order
of the atomic positions in the input file.