Rename the repository to aiida-quantumespresso-hp
The code that this plugin serves, the Hubbard module for
Quantum ESPRESSO, recently changed its name from Uscf to
HP, in preparation of adding support for the calculation
of Hubbard V parameters in addition to the Hubbard U.
Since this plugin at this point is not yet in use by any
other person than the developer, it is acceptable to change
the name of the plugin as well, for the sake of consistency.

The new binary of the Hubbard module is labeled hp.x and
the common acronym of the code is HP, which stand for
Hubbard Parameters. In line of that name change, we change
all the Calculation, Parser and WorkChain classes in this
plugin accordingly.