SelfConsistentHubbard: make the relax part truly optional
In the previous commit, an attempt was made to make the relaxing of the
structure during each iteration optional by providing the skip_relax
boolean input. However, the input group 'relax' was still required and
also no solution was provided for how to pass the required inputs
for the scf steps if the relax inputs where not specified.

The solution is to get rid of the skip_relax option and infer whether a
relaxation needs to be performed on whether the 'relax' input group is
specified. If not, then the 'scf' input group needs to be specified and
no relaxation will be performed.

This commit also ensure that the correct inputs are used for each run
pw related sub workchain and that changes to for example the parameters
input node in the scf calculations, are done to local deepcopies, making
sure that those changes are not included in the other workchains calls.