SelfConsistentHubbard: set stricter conv_thr in scf if not specified
For the scf calculations before the HpWorkChain, the conv_thr needs
to be set relatively strict to provide a reasonably well converged
charge density for the HpCalculation. In the case of the magnetic
insulator, the scf run with smearing can have a bit more relaxed
threshold, but for the fixed occupations followup it should be set
back to a tight value. Again all of these changes will only be done
if the user does not specify a value themselves for conv_thr