Add utils for dealing with kind order restriction of hp.x
The hp.x code requires that the in the atomic positions card of the
parent pw.x calculation are ordered in such a way that the Hubbard
kinds are listed first, or the code will crash. We are adding two
validation functions that will verify a PwCalculation or StructureData
to be a valid input for a HpCalculation by comparing the order of the
kinds to the list of Hubbard kinds.

Should the StructureData be incompatible, we provide a workfunction
to create a clone structure but with the kinds in the right order
according to a list of desired Hubbard kinds