Update workflows to newest features in aiida-core
The workflows now properly leverage the addition of portnamespaces
and the input and output exposing that comes with it. This vastly
simplifies the implementation of the wrapping workchains. In addition
the console scripts are updated to provide the necessary inputs such
as starting Hubbard U parameters, starting magnetizations and cutoffs
for the charge density and plane waves.