`HpCalculation`: allow `FolderData` as type for `parent_folder`
The `hp.x` code supports to parallelize calculations over sites that are
to be perturbed. A final calculation is then performed on the collected
results of the individual calculations. The intermediate results can be
collected in a `FolderData` which then becomes the `parent_folder` input
for a final `collect_chi` calculation.