`HpParser`: add support for "incomplete" `HpCalculation` runs
There are various legitimate `HpCalculation` runs that will not contain
the hubbard, chi and hubbard parameters output files. For example, in an
initialization run where only the symmetry analysis is performed to
determine which sites will have to be perturbed. Or a calculation where
not all perturbations are computed. In that case it is also not possible
for the final matrices to be computed. In these cases the parser should
not fail the calculation if the corresponding files are not present in
the retrieved folder.