`HpBaseWorkChain`: add handler for `ERROR_CONVERGENCE_NOT_REACHED` (#1)
The `hp.x` code currently does not support to restart from a partially
converged calculation. Therefore, the only chance of converging a failed
calculation is to restart it with different parameters. Specifically,
one can increase `niter_max` and reduce `alpha_mix` to increase the
chances of convergence being reached.

If these two input parameters are defined explicitly in the input
parameters, they are doubled and halved, respectively. Otherwise a hard
coded default is used that is higher or lower than the code's default,
respectively.