`HpCalculation`: correct and improve handling of parent folders
A `HpCalculation` always needs the outputs of a `PwCalculation`, which
was passed in through the `parent_folder` input. However, in certain
cases a `HpCalculation` depends on more than just this parent folder.
For example, when the calculation is parallelized over qpoints and/or
over atoms, at the end a final `HpCalculation` should be run that
collects the results of the individual computed perturbations and
performs the final computation of the chi matrix. This means that all
the outputs of the preceding calculation become additional "parent"
calculations.

To accommodate this, the input namespace `parent_hp` is added. This
namespace accepts `FolderData` nodes that are the `retrieved` output
nodes of a `HpCalculation` that computed part of the qpoint and
perturbed atomic space. Since for now the only reason for passing the
`parent_hp` input is to perform a chi computation, this flag is
automatically set in the inputs.

A `HpCalculation` needs the following files of the `PwCalculation` in
order to run:

  * `out/aiida.xml`
  * `out/aiida.save/data-file-schema.xml`
  * `out/aiida.save/*.UPF`
  * `out/aiida.save/charge-density.dat`
  * `out/aiida.save/occup.txt`
  * `out/aiida.save/wfc*.dat`

For simplicity's sake, the entire `out` folder of the `RemoteData`
folder of the `PwCalculation` are copied on the remote through the
`remote_copy_list`. This may be changed into a symlink in the future
for efficiency.

For the chi computation, the following additional files are required of
each of the parallelized `HpCalculation`s:

  * out/HP/aiida.chi.pert*
  * out/HP/aiida.dns.pert*

Since the number of files depends on the exactly how many qpoints and or
perturbed atoms were calculated, we have to resort to using wildcards to
copy those files. Since the `remote_copy_list` only supports exact
paths, instead we opt to always retrieve these files at the end of a
calculation. Since these files are not too big, they should not weigh
too heavily on the file repository. By doing this, they can be easily
copied over to the working directory of the collection calculation by
copying them to the scratch folder. To prevent the files from being
stored again in the repository of the `CalcJobNode`, the filepaths are
all added to the `provenance_exclude_list`.