`HpParallelizeAtomsWorkChain`: simplify parallelization mechanism
Since the `HpCalculation` was updated to always retrieve the
perturbation files of calculations, the mechanism of copying over all
required files to the working directory of the final chi computing
calculation is now a lot simpler. The `retrieved` folders of the
atomic calculations can now simply be passed into the `parent_hp` input
namespace. This means that the `collect_atomic_calculations` calc
function is superfluous and the intermediate `FolderData` that used to
serve as the `parent_folder` is no longer necessary.

This not only saves unnecessary copying back and forth of the output
files of the `PwCalculation` from the remote to the local machine, but
also removes unnecessary load on the local file repository.