`SelfConsistentHubbard`: convergence on per-site basis not per-kind
The original implementation of the workchain based convergence by
comparing the computed Hubbard U values per kind. However, `hp.x` can
determine new Hubbard sites, based on symmetry and occupations and so
really one should compare the values per site. A new calcfunction is
introduced `structure_relabel_kinds` that based on the `hubbard` output
dictionary of a `HpCalculation` will create a clone of the original
structure but with the kinds relabeled based on the analysis of the
calculation. It will also return a dictionary with the new Hubbard U
values based on the new kind labeling.