`SelfConsistentHubbardWorkChain`: fix bug with Hubbard site order
The `validate_inputs` method will validate the `StructureData` and make
sure that the Hubbard sites, as defined in the `hubbard_u` input, come
first in the structure definition. This is necessary because otherwise
the `HpCalculation` will fail. The check reordered the atomic sites if
necessary except didn't assign the corrected structure to the
`current_structure` variable in the context causing the workchain to
still use the original incorrect structure despite the check.