`HpCalculation`: automatically mark invalidates cache
By convention, exit codes with a status smaller than 400, are considered
to be unrecoverable errors. This means that these calculations should
not be considered as valid sources by the caching mechanism. This can be
signalled by setting `invalidate_cache=True` when declaring the exit
code. Instead of doing this manually for every port, since we have a
clear rule that all exit codes below 400 should have this value, this is
done in one place by using a subclass of `CalcJob` that has a subclass
of the `CalcJobProcessSpec` that automatically flips this switch based
on the exit status.

This trick is already implemented in `aiida-quantumespresso`, so the
only thing we have to do here is to subclass from the subclass
`aiida_quantumespresso.calculations.CalcJob` instead of the base class
`aiida.engine.Calcjob` that ships with `aiida-core`.