HpCalculation: exit code and handler for cholesky
The infamous Cholesky factorization error is also
possible in `hp.x`. Here we add the exit code, the
parser and the handler for the `HpBaseWorkChain`.
A riorganization of the `hp/parse_raw` is made to
make it easier for the future implementating new
exit codes, on the same line of aiida-qe.

On the contrary of `pw.x`, other diagonalization
options are not available. As from experience,
this error can come in magnetic system due to
numerical noise on parallel diagonalization. As such,
the implemented handler will set the diagonalization
flag to do that in serial, in order to avoid such noise.
This approach worked in many cases in my personal
experience, and it's the best we have at the moment.