Controlled submission of concurrent `HpBaseWorkChain`s
The workchains parallelizing atoms and q-points could not control 
the amount of concurrent sub-processes, specifically `HpBaseWorkChain`.
The user can now optionally control the maximum number of simultaneously
submitted `HpBaseWorkChains`. 

This is extremely useful in case one parallelizes over atoms and q-points,
giving more control to the user on the resources utilization.