Add `init_walltime` input to parallelize WorkChains (#62)
The `HpParallelizeQpointsWorkChain` and the  `HpParallelizeAtomsWorkChain` had an hard-coded
walltime for the  `HpBaseWorkChain` used for initialization, set to 3600 seconds. This might cause
some problems on certain HPC partitions (e.g. debug, tests, ...) where the maximum walltime can
lower than 1 hour. To make this more flexible, a new **non-db** input is added to both WorkChains 
with a default of 1 hour.