Fix the convergence check in self-consistent workflow
The convergence check on parameters had a bug in the logic.
That logic would work for 1 Hubbard atom only, and that's the
reason why this passed unseen in the tests. The fix is rather
simple, and the new tests are tested again a case which made
the old tests to make.