Change nearest-neighbour analysis and parser
The self-consistent workflow now checks the maximum number of neighbours
and defines the hp inputs so to print all the possible intersites couples. The
parser is then instructed to only keep the interactions that were defined in
the initial HubbardStructureData. This is useful in case the actual first neighbours
are different then the activated intersite interactions, thus a screening
of the HUBBARD.dat file cannot be avoided. This methodology allows for
having a more consistent definition and parsing of the nearest neighbours
using Voronoi tessellation.