Feature: add nearest neighbours analysis (#53)
Some structures may have onsite atoms of the same specie but with 
different nearest neighbours (e.g. Fe3O4). The hp.x code can handle 
this only by specifying a maximum radius, within which one can find 
the desired neighbours. Still, this is not yet sufficient, as the radius 
might change during relaxation. To solve this:

1. The self-consistent workflow now checks the maximum number of 
nearest neighbours using a Voronoi alrogithrm, as implemented in
Pymatgen, and defines the hp.x inputs so to print all the possible
intersites couples up to a maximum radius. This radius corresponds
to the maximum radius where the Voronoi algorithm looks for.

2. The parser is then instructed to only keep the interactions that were
defined in the initial HubbardStructureData. This is useful in case 
the actual first neighbours are different then the "activated" intersite
interactions, and the HUBBARD.dat cannot be parsed blindly.
This methodology allows for having a more consistent definition
and parsing of the nearest neighbours using Voronoi tessellation.